Organized_Schedules/requirements.txt
pandas>=1.5.0
python-dateutil>=2.8.2
supabase>=1.0.3
python-dotenv>=1.0.0
uuid>=1.30 